+++ src/database/db.py
@@ -1,5 +1,5 @@
-from bot.config import settings
-from bot.database.database_connector import DatabaseConnector
+from config import settings
+from database.database_connector import DatabaseConnector
 
 # todo: move from global scope
 # could open connection from init, which is bad
