+++ src/database/models/complete_lesson.py
@@ -1,7 +1,7 @@
 from sqlalchemy import ForeignKey
 from sqlalchemy.orm import Mapped, mapped_column
 
-from bot.database.models.base import Base
+from database.models.base import Base
 
 
 class CompleteLesson(Base):
