+++ src/database/models/session.py
@@ -1,8 +1,8 @@
 from sqlalchemy import ForeignKey
 from sqlalchemy.orm import Mapped, mapped_column
 
-from bot.database.models.base import Base
 from bot.resources.enums import SessionStartsFrom, SessionStatus
+from database.models.base import Base
 
 
 class Session(Base):
