+++ src/database/models/sticker.py
@@ -1,7 +1,7 @@
 from sqlalchemy.orm import Mapped, mapped_column
 
-from bot.database.models.base import Base
 from bot.resources.enums import StickerType
+from database.models.base import Base
 
 
 class Sticker(Base):
